+++ packages/swirl-components/src/components/swirl-avatar-group/swirl-avatar-group.stories.ts
@@ -14,6 +14,7 @@ export default {
     },
   },
   component: "swirl-avatar-group",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,