+++ packages/swirl-components/src/components/swirl-button-group/swirl-button-group.stories.ts
@@ -9,6 +9,7 @@ export default {
     },
   },
   component: "swirl-button-group",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,