+++ packages/swirl-components/src/components/swirl-pdf-reader/swirl-pdf-reader.stories.ts
@@ -3,6 +3,7 @@ import Docs from "./swirl-pdf-reader.mdx";
 
 export default {
   component: "swirl-pdf-reader",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,