+++ packages/swirl-components/src/components/swirl-separator/swirl-separator.spec.tsx
@@ -10,7 +10,7 @@ describe("swirl-separator", () => {
     });
 
     expect(page.root).toEqualHtml(`
-      <swirl-separator class="separator" role="separator" style="padding-top: var(--s-space-8); padding-bottom: var(--s-space-8);">
+      <swirl-separator aria-orientation="horizontal" class="separator separator--orientation-horizontal" role="separator" style="padding-top: var(--s-space-8); padding-bottom: var(--s-space-8);">
         <mock:shadow-root>
           <span class="separator__line"></span>
         </mock:shadow-root>