+++ packages/swirl-components/src/components/swirl-tab-bar/swirl-tab-bar.stories.ts
@@ -3,6 +3,7 @@ import Docs from "./swirl-tab-bar.mdx";
 
 export default {
   component: "swirl-tab-bar",
+  tags: ["autodocs"],
   parameters: {
     docs: {
       page: Docs,